feat: 'Pressure for the last 30 days' pág 48
Continuación del anterior commit 'Implementing the UI mockup':
Making the homepage into something resembling my UI mockup required several steps:3. Adding an API to get blood-pressure readings for the last 30 days.

- Mejora: cada usuario con rol USER sólo puede ver y manipular sus datos, ADMIN puede verlos todos excepto las Preferences.

- Mejora: la opción de DELETE sólo recibe el ID, antes recibía la entidad completa haciendo una consulta GET al servidor.

- Curiosidad: hay un PreferenceDTO con su PreferenceMapper con su anotación @Mapper de org.mapstruct, pero ejecutando la app desde ECLIPSE no lo inyecta correctamente, desde IntelliJ IDEA si. Lo cambié por un Mapper personalizado, sin mapstruct.

- Mejora: en nvd3 para que funcione bien 'useInteractiveGuideline: true' los datos del eje x se deben cargar de menos a mas (http://stackoverflow.com/questions/22259025/nvd3-chart-useinteractiveguidelinetrue-not-updating-properly).

- Mejora: nvd3 - traducción de las etiquetas.

- Nota: nvd3 - se pueden ver todas las propiedades en http://krispo.github.io/angular-nvd3/#/lineChart con la opción "Extended".

+++ src/main/java/com/desprogramar/jhipster/web/rest/BloodPressureResource.java
@@ -3,11 +3,16 @@ package com.desprogramar.jhipster.web.rest;
 import com.codahale.metrics.annotation.Timed;
 import com.desprogramar.jhipster.domain.BloodPressure;
 import com.desprogramar.jhipster.repository.BloodPressureRepository;
+import com.desprogramar.jhipster.repository.UserRepository;
 import com.desprogramar.jhipster.repository.search.BloodPressureSearchRepository;
+import com.desprogramar.jhipster.security.AuthoritiesConstants;
+import com.desprogramar.jhipster.security.SecurityUtils;
+import com.desprogramar.jhipster.web.rest.dto.BloodPressureByPeriod;
 import com.desprogramar.jhipster.web.rest.util.HeaderUtil;
 import com.desprogramar.jhipster.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
@@ -20,8 +25,8 @@ import javax.inject.Inject;
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.time.ZonedDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -42,6 +47,9 @@ public class BloodPressureResource {
     @Inject
     private BloodPressureSearchRepository bloodPressureSearchRepository;
     
+    @Autowired
+    private UserRepository userRepository;
+    
     /**
      * POST  /bloodPressures -> Create a new bloodPressure.
      */
@@ -53,6 +61,10 @@ public class BloodPressureResource {
         log.debug("REST request to save BloodPressure : {}", bloodPressure);
         if (bloodPressure.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("bloodPressure", "idexists", "A new bloodPressure cannot already have an ID")).body(null);
+        }
+        if (bloodPressure.getUser() == null) {
+            log.debug("No user passed in, using current user");
+            bloodPressure.setUser(userRepository.findOneByUserIsCurrentUser().get());
         }
         BloodPressure result = bloodPressureRepository.save(bloodPressure);
         bloodPressureSearchRepository.save(result);
@@ -73,6 +85,10 @@ public class BloodPressureResource {
         if (bloodPressure.getId() == null) {
             return createBloodPressure(bloodPressure);
         }
+        if (bloodPressure.getUser() == null) {
+            log.debug("No user passed in, using current user");
+            bloodPressure.setUser(userRepository.findOneByUserIsCurrentUser().get());
+        }
         BloodPressure result = bloodPressureRepository.save(bloodPressure);
         bloodPressureSearchRepository.save(result);
         return ResponseEntity.ok()
@@ -90,7 +106,12 @@ public class BloodPressureResource {
     public ResponseEntity<List<BloodPressure>> getAllBloodPressures(Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to get a page of BloodPressures");
-        Page<BloodPressure> page = bloodPressureRepository.findAll(pageable); 
+        Page<BloodPressure> page;
+        if (SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+            page = bloodPressureRepository.findAll(pageable);
+        } else {
+            page = bloodPressureRepository.findAllByUserIsCurrentUser(pageable);
+        }
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/bloodPressures");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
@@ -105,11 +126,13 @@ public class BloodPressureResource {
     public ResponseEntity<BloodPressure> getBloodPressure(@PathVariable Long id) {
         log.debug("REST request to get BloodPressure : {}", id);
         BloodPressure bloodPressure = bloodPressureRepository.findOne(id);
-        return Optional.ofNullable(bloodPressure)
-            .map(result -> new ResponseEntity<>(
-                result,
-                HttpStatus.OK))
-            .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        if (bloodPressure == null || (
+        		!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+            	!bloodPressure.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) ) {
+//            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(id.toString(), "idNoExists", "")).body(null);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    	return  new ResponseEntity<>(bloodPressure, HttpStatus.OK);
     }
 
     /**
@@ -121,6 +144,13 @@ public class BloodPressureResource {
     @Timed
     public ResponseEntity<Void> deleteBloodPressure(@PathVariable Long id) {
         log.debug("REST request to delete BloodPressure : {}", id);
+        BloodPressure bloodPressure = bloodPressureRepository.findOne(id);
+        if (bloodPressure != null &&
+        		!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+        		!bloodPressure.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) {
+//        	return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("bloodPressure", "idNoExiste", "El ID '"+id+"' no existe!")).build();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         bloodPressureRepository.delete(id);
         bloodPressureSearchRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("bloodPressure", id.toString())).build();
@@ -140,4 +170,22 @@ public class BloodPressureResource {
             .stream(bloodPressureSearchRepository.search(queryStringQuery(query)).spliterator(), false)
             .collect(Collectors.toList());
     }
+    
+    /**
+     * GET  /bp-by-days/:days -> get the "days" del período.
+     */
+    @RequestMapping(value = "/bp-by-days/{days}",
+        method = RequestMethod.GET,
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    @Timed
+    public ResponseEntity<BloodPressureByPeriod> getByDays(@PathVariable Integer days) {
+        log.debug("REST request to get BloodPressure del período de días: {}", days);
+    	ZonedDateTime now = ZonedDateTime.now().withNano(0);
+    	ZonedDateTime desde = now.minusDays(days).withHour(0).withMinute(0).withSecond(0);
+    	
+    	List<BloodPressure> bloodList = bloodPressureRepository.findByTimestampAfterCurrentUserOrderByTime(desde, now);
+    	BloodPressureByPeriod lastDaysDTO = new BloodPressureByPeriod(days, bloodList);
+    	return new ResponseEntity<>(lastDaysDTO, HttpStatus.OK);
+    }
+
 }
