+++ src/main/java/com/desprogramar/jhipster/web/rest/WeightResource.java
@@ -2,12 +2,16 @@ package com.desprogramar.jhipster.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
 import com.desprogramar.jhipster.domain.Weight;
+import com.desprogramar.jhipster.repository.UserRepository;
 import com.desprogramar.jhipster.repository.WeightRepository;
 import com.desprogramar.jhipster.repository.search.WeightSearchRepository;
+import com.desprogramar.jhipster.security.AuthoritiesConstants;
+import com.desprogramar.jhipster.security.SecurityUtils;
 import com.desprogramar.jhipster.web.rest.util.HeaderUtil;
 import com.desprogramar.jhipster.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
@@ -21,7 +25,6 @@ import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
@@ -42,6 +45,9 @@ public class WeightResource {
     @Inject
     private WeightSearchRepository weightSearchRepository;
     
+    @Autowired
+    private UserRepository userRepository;
+
     /**
      * POST  /weights -> Create a new weight.
      */
@@ -53,6 +59,10 @@ public class WeightResource {
         log.debug("REST request to save Weight : {}", weight);
         if (weight.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("weight", "idexists", "A new weight cannot already have an ID")).body(null);
+        }
+        if(weight.getUser() == null) {
+            log.debug("No user passed in, using current user");
+            weight.setUser(userRepository.findOneByUserIsCurrentUser().get());
         }
         Weight result = weightRepository.save(weight);
         weightSearchRepository.save(result);
@@ -73,6 +83,10 @@ public class WeightResource {
         if (weight.getId() == null) {
             return createWeight(weight);
         }
+        if(weight.getUser() == null) {
+            log.debug("No user passed in, using current user");
+            weight.setUser(userRepository.findOneByUserIsCurrentUser().get());
+        }
         Weight result = weightRepository.save(weight);
         weightSearchRepository.save(result);
         return ResponseEntity.ok()
@@ -90,7 +104,12 @@ public class WeightResource {
     public ResponseEntity<List<Weight>> getAllWeights(Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to get a page of Weights");
-        Page<Weight> page = weightRepository.findAll(pageable); 
+        Page<Weight> page;
+        if (SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+            page = weightRepository.findAll(pageable);
+        } else {
+            page = weightRepository.findAllByUserIsCurrentUser(pageable);
+        }
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/weights");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
@@ -105,11 +124,12 @@ public class WeightResource {
     public ResponseEntity<Weight> getWeight(@PathVariable Long id) {
         log.debug("REST request to get Weight : {}", id);
         Weight weight = weightRepository.findOne(id);
-        return Optional.ofNullable(weight)
-            .map(result -> new ResponseEntity<>(
-                result,
-                HttpStatus.OK))
-            .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        if (weight == null || (
+        		!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+        		!weight.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) ) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    	return  new ResponseEntity<>(weight, HttpStatus.OK);
     }
 
     /**
@@ -121,6 +141,12 @@ public class WeightResource {
     @Timed
     public ResponseEntity<Void> deleteWeight(@PathVariable Long id) {
         log.debug("REST request to delete Weight : {}", id);
+        Weight weight = weightRepository.findOne(id);
+        if (weight == null || (
+        		!SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN) &&
+        		!weight.getUser().getLogin().equals(SecurityUtils.getCurrentUserLogin())) ) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         weightRepository.delete(id);
         weightSearchRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("weight", id.toString())).build();
